9512: drying up test [skip ci]
