+++ web-client/integration-tests/journey/docketClerkAddsPretrialMemorandumToCase.js
@@ -1,0 +1,86 @@
+import { OBJECTIONS_OPTIONS_MAP } from '../../../shared/src/business/entities/EntityConstants';
+import { contactPrimaryFromState, fakeFile } from '../helpers';
+
+export const docketClerkAddsPretrialMemorandumToCase = (
+  cerebralTest,
+  { caseNumber, filedByPetitioner, filedByPractitioner },
+) => {
+  return it('Docket clerk adds pretrial memorandum to case', async () => {
+    await cerebralTest.runSequence('gotoCaseDetailSequence', {
+      docketNumber: cerebralTest[`docketNumber${caseNumber}`],
+    });
+
+    await cerebralTest.runSequence('gotoAddPaperFilingSequence', {
+      docketNumber: cerebralTest[`docketNumber${caseNumber}`],
+    });
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'dateReceivedMonth',
+      value: 4,
+    });
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'dateReceivedDay',
+      value: 30,
+    });
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'dateReceivedYear',
+      value: 2001,
+    });
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'primaryDocumentFile',
+      value: fakeFile,
+    });
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'primaryDocumentFileSize',
+      value: 100,
+    });
+
+    if (filedByPetitioner) {
+      const contactPrimary = contactPrimaryFromState(cerebralTest);
+
+      await cerebralTest.runSequence(
+        'updateFileDocumentWizardFormValueSequence',
+        {
+          key: `filersMap.${contactPrimary.contactId}`,
+          value: true,
+        },
+      );
+    }
+
+    if (filedByPractitioner) {
+      await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+        key: 'partyIrsPractitioner',
+        value: true,
+      });
+    }
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'eventCode',
+      value: 'PMT',
+    });
+
+    await cerebralTest.runSequence('updateDocketEntryFormValueSequence', {
+      key: 'objections',
+      value: OBJECTIONS_OPTIONS_MAP.NO,
+    });
+
+    await cerebralTest.runSequence('submitPaperFilingSequence');
+
+    expect(cerebralTest.getState('validationErrors')).toEqual({});
+
+    expect(cerebralTest.getState('alertSuccess').message).toEqual(
+      'Your entry has been added to docket record.',
+    );
+
+    expect(cerebralTest.getState('currentPage')).toEqual('CaseDetailInternal');
+    expect(cerebralTest.getState('form')).toEqual({});
+
+    cerebralTest.docketEntryId = cerebralTest
+      .getState('caseDetail.docketEntries')
+      .find(doc => doc.eventCode === 'PMT').docketEntryId;
+  });
+};
